Move AnalysisResults to a separate file
