+++ examples/general.py
@@ -63,7 +63,7 @@ if __name__ == "__main__":
                                              {'shift':shift_GHz, 'shift_units': 'GHz',
                                              'width': width_GHz, 'width_units': 'Hz'},
                                              name = 'test1_analysis',
-                                             fit_model=brim.Data.AnalysisResults.FitModel.Lorentzian)
+                                             fit_model=brim.AnalysisResults.FitModel.Lorentzian)
     f.close()
 
 
@@ -108,13 +108,13 @@ if __name__ == "__main__":
     pt = ar.list_existing_peak_types()
     qt = ar.list_existing_quantities()
     # get the image of the shift quantity for the average of the Stokes and anti-Stokes peaks
-    img, px_size = ar.get_image(brim.Data.AnalysisResults.Quantity.Shift, brim.Data.AnalysisResults.PeakType.average)
+    img, px_size = ar.get_image(brim.AnalysisResults.Quantity.Shift, brim.AnalysisResults.PeakType.average)
     # get the units of the shift quantity
-    u = ar.get_units(brim.Data.AnalysisResults.Quantity.Shift)
+    u = ar.get_units(brim.AnalysisResults.Quantity.Shift)
 
     # get a quantity at a specific pixel (coord) in the image
     coord = (1,3,4)
-    qt_at_px = ar.get_quantity_at_pixel(coord, brim.Data.AnalysisResults.Quantity.Shift, brim.Data.AnalysisResults.PeakType.average)
+    qt_at_px = ar.get_quantity_at_pixel(coord, brim.AnalysisResults.Quantity.Shift, brim.AnalysisResults.PeakType.average)
     assert img[coord]==qt_at_px
 
     # get the spectrum in the image at a specific pixel (coord)
